requirements.txt
# LangChain
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-core>=0.1.0

# CrewAI
crewai>=0.28.0
crewai-tools>=0.2.0

# AutoGen
pyautogen>=0.2.0

# Common dependencies
openai>=1.10.0
requests>=2.31.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4

# Database connectors (for testing)
psycopg2-binary>=2.9.9
boto3>=1.34.0
